// frontend/src/generated/api/models/SensitivityMode.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EnergyForecast API
 * Energy consumption analysis and forecasting API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SensitivityMode = {
    CompleteOnly: 'complete_only',
    Coverage90: 'coverage_90'
} as const;
export type SensitivityMode = typeof SensitivityMode[keyof typeof SensitivityMode];


export function instanceOfSensitivityMode(value: any): boolean {
    for (const key in SensitivityMode) {
        if (Object.prototype.hasOwnProperty.call(SensitivityMode, key)) {
            if (SensitivityMode[key as keyof typeof SensitivityMode] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SensitivityModeFromJSON(json: any): SensitivityMode {
    return SensitivityModeFromJSONTyped(json, false);
}

export function SensitivityModeFromJSONTyped(json: any, ignoreDiscriminator: boolean): SensitivityMode {
    return json as SensitivityMode;
}

export function SensitivityModeToJSON(value?: SensitivityMode | null): any {
    return value as any;
}

export function SensitivityModeToJSONTyped(value: any, ignoreDiscriminator: boolean): SensitivityMode {
    return value as SensitivityMode;
}